Bugfixes searchalgorithm

+++ backend/APIApplication/DatabaseLibrary/IDbSearch.cs
@@ -4,7 +4,7 @@ namespace DatabaseLibrary;
 
 public interface IDbSearch
 {
-    List<int> GetStoresInRange(double x, double y, int range);
+    List<int> GetStoresInRange(double x, double y, double range);
     List<Store> GetDataFromStores(List<int> topStores);
     List<Product> GetProductsFromSpecificStores(List<int> storeKeys, string productName, string measurement);
 }
